op3: Add shim to make front camera not take 4 MP pics in bright light
After adding the previous shims for the front camera, the front camera no
longer crashes when taking pictures outdoors/under bright light, but the
resolution of these photos is incorrect. Attempting to take a picture with
the front camera under bright light yields a 4 MP photo (2304x1728) instead
of the user-requested resolution (e.g. 16 MP).

In order to fix this, the handleSuperResoultion and isSuperResoultion
functions are shimmed. However, shimming these to always return true breaks
photo capture for the rear camera, so make these functions only return true
when the front camera is in use.

These hacks are only for the OnePlus 3T and would cause breakage on the
OnePlus 3 if they are enabled on it.

Change-Id: I77ca1875935b6b7789daa287c796567041d33e29